Integrate new deid chart with the alvearie-ingestion chart
This commit includes the deid chart created in a previous commit with
the alvearie-ingestion chart. By default the deid flow services
(secondary FHIR and deid service) are enabled, but creating a second
values.yaml file with them disabled would be trivial, or we could add
the one-lines instructions to deploy with deid off.

This commit also includes a default _helpers.tpl that will likely be
necessary for a future task on writing the NOTES.txt file for the
alvearie  chart.